pyproject.toml
[project]
name = "starphone-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Giovanni Padilha", email = "giovanni.padilha@outlook.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.117.1",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sqlmodel>=0.0.25",
]

#[project.scripts]
#starphone-api = "starphone_api.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.5.0",
    "ruff>=0.13.1",
]
